pyproject.toml
[tool.poetry]
name = "ing-pyt-bms-clean"
version = "0.1.0"
description = "Plantilla para crear un proyecto Python en arquitectura Clean"
authors = ["Tu Nombre <tu.email@ejemplo.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.100.0"
uvicorn = "^0.23.2"
httpx = "^0.27.0"
azure-keyvault-secrets = "4.7.0"
azure-identity = "1.13.0"
pydantic = "^2.0"
dependency-injector = "4.41.0"
azure-core = "1.28.0"
python-dotenv = "1.0.0"
tensorflow = "2.16.1"
tensorflow-io-gcs-filesystem = "0.31.0"
fastapi-mcp = "^0.3.7"
keras = "^3.0"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.0"
pytest-asyncio = "0.21.1"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.5.0"
coverage = "^7.9.1"
pytest-cov = "^6.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88